dmpy: validate dm_task data by task type
Keep track of the task type that a DmTask object is initialised
with, and use this to determine what data is present in the ioctl
response.

This avoids needing to copy libdm-private headers into dmpy, or
modify libdevmapper to expose this information.

A table of flags values, indexed by dm_task type is used to
initalise a DmTask object's ob_flags following a successful ioctl.