dmpy: add experimental DmStats region support
Add support for a DmStats object to act as an indexable container:
regions may be accessed by subscripting the container:

  >>> dms.list()
  >>> len(dms)
  4
  >>> dms[0].nr_areas()  # Yields a DmStatsRegion object.
  8
  >>>

The DmStats object maintains a cache of active DmStatsRegions,
as an array of PyObject pointers. To avoid cyclical garbage
collection problems, the array is populated with weak reference
objects that reference the contained DmStatsRegion. This allows
effective collection of an unreferenced DmStatsRegion object
while protecting it from having its parent DmStats object
deallocated while it is still in use: keeping a reference from
the DmStatsRegion to the DmStats is not an option since the
reference cycle will prevent the DmStats object from ever being
cleaned up (leading to libdevmapper pool leak warnings at exit).