dmpy: add DmStatsRegion class
Add a base class to represent a region of a DmStats object. Since
the child DmStatsRegion object shares its state (via the dm_stats
handle) with the parent, it must maintain a strong reference on
the parent to prevent destruction of the object and its dm_stats
handle. Normally the lifespan of child objects is shorter than
that of the parent so the main object will be reaped soon after
the final external reference is dropped.

To prevent cyclical references that would prevent or delay garbage
collection (possibly leading to libdevmapper pool leak warnings
when shutting down the interpreter), the containing objects in a
DmStats object hierarchy can only ever maintain weak references to
child objects. To facilitate this scheme the DmStatsRegion
implememts Python Weakref, allowing a containing DmStats to
maintain a cache of previously seen region objects without
introducing reference cycles.