dmpy: make DmStats a container type
Implement Sequence methods for the DmStats object: this allows the
use of array subscript notation and 'for x in y' syntax to retrieve
or iterate over the regions contained in it.

The DmStats.__getitem__ method instantiates new DmStatsRegion
objects on demand, and maintains a weak cache of previously
accessed items. If the referent of a cached reference has expired,
it is treated as a cache miss and a newly created object is
inserted into the cache in its place.