dmpy: make DmStatsRegion a container of DmStatsArea objects
Now that the top-level region indexing (through DmStats[r_id]) is
complete, add the next level down: implement Sequence methods for
DmStatsRegion to allow indexing to the area_id level, and add a
new DmStatsArea class to represent the counter data contained in
a DmStatsRegion.

The caching scheme used in DmStats is extended into DmStatsRegion:
a DmStatsRegion maintains a cache of weak references to DmStatsArea
objects (indexed by area_id) that were previously instantiated from
this DmStatsRegion, and creates new DmStatsArea objects in response
to a cache miss on an area_id lookup.

As before, all data objects that rely on the state of the parent
DmStats object maintain a strong reference to it, and store a
pointer to the corresponding DmStatsObject.