dmpy: protect DmStats->ob_dms by sequence number
By maintaining strong references from child objects of DmStats to
the parent, we prevent destruction of the parent while the children
are still referenced, however, the stored dm_stats pointer is still
mutable, and may be changed in a manner which invalidates all child
objects that were previously referencing it.

In practice, this will occur with any sequence of operations that
involve accessing a child object by a reference stored prior to a
bind, list, or populate operation on the same containing DmStats:

  >>> dms.list()
  >>> region = dms[0] # obtain reference to DmStatsRegion
  >>> dms.list()      # region is now invalidated

To prevent this, store a sequence number with the DmStats object
that is initialised to zero, and that increments each time that a
handle invalidating operation occurs.

Child objects that need to know whether the handle is stale (from
their point of view) or not should store the sequence number that
was current at the time of their initialisation, and compare this
to the current sequence number in the parent DmStats object: if
the two do not match the child cannot use the parent reference,
and must raise the LookupError exception on any operation that
would access the parent state.