dmpy: add DmStatsArea.get_attr() for counter values
Expose dm_stats area counters via named attributes looked up from
a tp_getattro method. Attribute names are checked against the list
of counters and if a match is found, the corresponding value is
returned as a Python int. If no match exists the method call is
is forwarded to PyObject_GenericGetAttr() for normal resolution
via __dict__ etc.

E.g.:

>>> dms.populate("", region_id=dm.STATS_REGIONS_ALL)
<dmpy.DmStats object at 0x7fc74be55708>
>>> dms[0][0].READS_COUNT
67
>>> dms[0][0].WRITES_COUNT
23002